refactoring(15): decompose WorkflowPage.tsx (extract workflowNodes)
Second slice of req-15 (multi-PR). WorkflowPage.tsx 1720 -> 968 lines: extract
the React Flow node/modal layer (skillColor + SKILL_PALETTE, PhaseNode/
ConditionNode/ParallelNode + nodeTypes, ConditionModal/SkillDrawer/NodeEditModal/
SkillModal, SkillMcpEntry, inputStyle) VERBATIM into
components/Workflow/workflowNodes.tsx; container imports skillColor/nodeTypes/
the four modals + SkillMcpEntry. Moved code byte-identical (modulo export kw);
removed the now-dead reactflow Handle/Position/NodeTypes container imports.

Checks: typecheck 0 non-test errors; build ok; vitest 4 passed / 2 failed
(pre-existing brittle PaperBench tests). req-15 file RETAINED (multi-PR, §9/§10).
Remaining: StepResources (has clean units), SettingsPage + DetailPanel
(monoliths needing higher-risk prop-contract refactoring).
