+++ ari-core/ari/viz/frontend/src/components/Workflow/README.md
@@ -6,4 +6,5 @@ Workflow page — displays the run's workflow stages/pipeline.
 
 - `README.md` — this file.
 - `index.ts` — barrel re-export.
+- `workflowNodes.tsx` — React Flow custom nodes + edit/skill/condition modals (extracted from WorkflowPage in req 15).
 - `WorkflowPage.tsx` — workflow view.
